% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geyser_hook.R
\name{geyser_hook}
\alias{geyser_hook}
\title{geyser hook}
\usage{
geyser_hook(path, package_name, ...)
}
\arguments{
\item{path}{Name of the folder to create the package in.
This will also be used as the package name.}

\item{package_name}{Package name to use. By default, {golem} uses
\code{basename(path)}. If \code{path == '.'} & \code{package_name} is
not explicitly set, then \code{basename(getwd())} will be used.}

\item{...}{Arguments passed from \code{create_golem()}, unused in the default
function.}
}
\value{
Used for side effect
}
\description{
This hook creates an Old Faithful Geyser Data
application. It is mainly inteded as a simple example of how to
use the \code{project_hook} argument of \code{create_golem()}.
}
\examples{
if (FALSE){
  pth <- tempfile(pattern = "geyser")
  golem::create_golem(pth, project_hook = geyser_hook)
  readLines(file.path(pth, "R/app_ui.R"))
  unlink(pth, recursive = TRUE)
}
}
